Change AnalysisTab to represent the object in the UI rather than a proxy object.
Previously, AnalysisTab only exists to provide the functions when an analysis option is executed.
The class did not represent anything physical in the UI; when finding the instance objects inside
the tab, it had to call self.parent().findChild, ie. let the main window find the object instead.

Now, it extends from QWidget, and represents the object itself. This allows the class to find its
children directly using self.findChild, and inherits everything that the UI object has, eg.
addWidget, which would physically change the UI as opposed to not working as it would before. This
is done by using promoted widgets in Qt Designer. Qt now looks for the module and evokes the code
when loading the object in the UI.

As a side effect, __init__ can no longer be used, because during uic.loadUi is processing the .ui
file, the class instance is initialised before any children are added to it + its parents are set.
An '_activate' method is used instead, which is called in the main class, after uic.loadUi finishes
its job.

To refer to the functions/widgets in the main window, the method .window() is used instead of
.parent().