Refactor file and class schema
uic.loadUi was discovered to automatically assign attributes, and as such, the findChild/findObjects methods
are considered not necessary in most circumstances. However, this means that a seperate .ui file must be
included for each composite widget (container).

Adds new classes/files DirectoryWidget and MediaWidget, previously in AnalysisMain

Rename "Continue" to "Analyse", "Filter function" to "Damping function" in autospec